requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==22.1.0
black==24.10.0
blessed==1.20.0
blinker==1.9.0
certifi==2022.9.24
cffi==1.17.0
cfgv==3.4.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.7
distlib==0.3.9
editor==1.6.6
execnet==2.1.0
exrex==0.11.0
Faker==25.1.0
filelock==3.13.1
Flask==3.1.0
fsspec==2024.10.0
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==2.6.3
idna==3.4
iniconfig==1.1.1
inquirer==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
logger-config==0.3
lxml==5.1.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==21.3
pathspec==0.12.1
platformdirs==4.3.6
playwright==1.43.0
pluggy==1.0.0
pre_commit==4.0.1
psycopg2-binary==2.9.10
py==1.11.0
PyAudio==0.2.14
pycparser==2.22
pydantic==1.10.2
pyee==11.1.0
pyparsing==3.0.9
pytest==7.1.3
pytest-base-url==2.1.0
pytest-playwright==0.4.4
pytest-repeat==0.9.1
pytest-rerunfailures==13.0
pytest-timeout==2.2.0
pytest-xdist==3.6.1
pytest-xvfb==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-guerrillamail==0.2.0
python-json-logger==2.0.7
python-slugify==6.1.2
PyVirtualDisplay==3.0
PyYAML==6.0.2
readchar==4.2.1
regex==2024.11.6
requests==2.31.0
rfc3986==1.5.0
runs==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
srt==3.5.3
structlog==21.5.0
sympy==1.13.1
text-unidecode==1.3
tiktoken==0.8.0
tomli==2.0.1
torch==2.5.1
tqdm==4.67.0
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.28.0
vosk==0.3.44
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
whisper==1.1.10
xmod==1.8.1
